requirements.txt
boto3==1.26.99
botocore==1.29.99
mysql-connector-python==8.0.32
numpy==1.23.5
pandas==1.5.3
Pillow==9.4.0
pydicom==2.3.1
s3fs==0.4.2
streamlit==1.20.0
git+https://github.com/glassonion1/anonypy.git#egg=anonypy
streamlit-aggrid==0.3.4.post2